Moves history to own section of history.
refs #2516
