+++ src/actionCreators/templates.js
@@ -2,7 +2,8 @@
 import { addError } from 'actions/errors'
 import { validateTemplates } from './templateValidationHelpers'
 import Config from 'Config'
-import { addTemplates, addTemplateHistoryByResult } from 'actions/templates'
+import { addTemplates } from 'actions/templates'
+import { addTemplateHistoryByResult } from 'actions/history'
 import { selectSubjectAndPropertyTemplates } from 'selectors/templates'
 import TemplatesBuilder from 'TemplatesBuilder'
 import { fetchResource } from 'sinopiaApi'
