+++ src/actions/templates.js
@@ -1,14 +1,6 @@
-export const addTemplateHistory = (resourceTemplate) => ({
-  type: 'ADD_TEMPLATE_HISTORY',
-  payload: resourceTemplate,
-})
-
-export const addTemplateHistoryByResult = (result) => ({
-  type: 'ADD_TEMPLATE_HISTORY_BY_RESULT',
-  payload: result,
-})
-
 export const addTemplates = (subjectTemplate) => ({
   type: 'ADD_TEMPLATES',
   payload: subjectTemplate,
 })
+
+export const noop = () => {}
