+++ src/reducers/index.js
@@ -27,9 +27,10 @@ import { showCopyNewMessage } from './messages'
 import {
   exportsReceived,
 } from './exports'
+import { addTemplates } from './templates'
 import {
-  addTemplates, addTemplateHistory, addTemplateHistoryByResult,
-} from './templates'
+  addTemplateHistory, addTemplateHistoryByResult,
+} from './history'
 import {
   clearSearchResults, setSearchResults,
 } from './search'
@@ -62,8 +63,6 @@ const appHandlers = {
 const editorHandlers = {
   ADD_ERROR: addError,
   ADD_MODAL_MESSAGE: addModalMessage,
-  ADD_TEMPLATE_HISTORY: addTemplateHistory,
-  ADD_TEMPLATE_HISTORY_BY_RESULT: addTemplateHistoryByResult,
   CLEAR_ERRORS: clearErrors,
   CLEAR_MODAL_MESSAGES: clearModalMessages,
   CLEAR_RESOURCE: clearResourceFromEditor,
@@ -100,6 +99,11 @@ const entityHandlers = {
   SHOW_PROPERTY: showProperty,
 }
 
+const historyHandlers = {
+  ADD_TEMPLATE_HISTORY: addTemplateHistory,
+  ADD_TEMPLATE_HISTORY_BY_RESULT: addTemplateHistoryByResult,
+}
+
 const searchHandlers = {
   CLEAR_SEARCH_RESULTS: clearSearchResults,
   SET_SEARCH_RESULTS: setSearchResults,
@@ -115,6 +119,7 @@ const appReducer = combineReducers({
   app: createReducer(appHandlers),
   editor: createReducer(editorHandlers),
   entities: createReducer(entityHandlers),
+  history: createReducer(historyHandlers),
   search: createReducer(searchHandlers),
 })
