+++ src/selectors/history.js
@@ -1,0 +1,3 @@
+export const selectHistoricalTemplates = (state) => state.history.templates
+
+export const noop = () => {}
