+++ src/store.js
@@ -20,7 +20,6 @@ export const initialState = {
     currentResource: undefined,
     currentComponent: {},
     errors: {}, // {<error key>: [errors...]} or {<error key>: {<resourceKey>: [errors...]}}
-    historicalTemplates: [],
     lastSave: {}, // {<resourceKey>: date}
     modal: {
       name: undefined, // Name of modal to show. Should only be one at a time.
@@ -40,6 +39,10 @@ export const initialState = {
     subjects: {},
     subjectTemplates: {},
     values: {},
+  },
+  history: {
+    templates: [],
+    search: [],
   },
   search: {
     // Search model:
